[PDS-189] refactor: 자산 정보 확인 이메일 알림 로직 이동
