+++ src/main/java/com/example/pladialmserver/notification/service/PushNotificationService.java
@@ -3,9 +3,11 @@ package com.example.pladialmserver.notification.service;
 import com.example.pladialmserver.global.Constants;
 import com.example.pladialmserver.global.exception.BaseException;
 import com.example.pladialmserver.global.exception.BaseResponseCode;
+import com.example.pladialmserver.global.utils.EmailUtil;
 import com.example.pladialmserver.notification.dto.FcmMessage;
 import com.example.pladialmserver.notification.entity.PushNotification;
 import com.example.pladialmserver.notification.repository.PushNotificationRepository;
+import com.example.pladialmserver.user.dto.request.SendAssetsEmailReq;
 import com.example.pladialmserver.user.entity.User;
 import com.example.pladialmserver.user.repository.user.UserRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -23,6 +25,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.io.IOException;
 import java.util.List;
 
+import static com.example.pladialmserver.global.Constants.EmailNotification.*;
+
 
 @Slf4j
 @Component
@@ -33,6 +37,7 @@ public class PushNotificationService {
     private final ObjectMapper objectMapper;
     private final PushNotificationRepository notificationRepository;
     private final UserRepository userRepository;
+    private final EmailUtil emailUtil;
 
     @Transactional
     public void sendNotification(String category, String type, User user) throws IOException {
@@ -64,6 +69,15 @@ public class PushNotificationService {
             } catch (IOException e) {
                 throw new BaseException(BaseResponseCode.NOTIFICATION_TOKEN_NOT_FOUND);
             }
+        });
+    }
+
+    // 자산 정보 확인 알림 / 메일 전송
+    @Scheduled(cron = "0 0 9 1 */3 ?", zone = "GMT+9:00")
+    public void sendAssetsEmail(){
+        userRepository.findByAssetsIsNotNull().forEach(user -> {
+            emailUtil.sendEmail(user.getEmail(), COMPANY_NAME + ASSETS_TITLE
+                    , emailUtil.createAssetsData(SendAssetsEmailReq.toDto(user)), ASSETS_TEMPLATE);
         });
     }
